# v1/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Pi-Hole",
    "description": "The Pi-Hole® is a DNS sinkhole that protects your devices from unwanted content, without installing any client-side software.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
    "/aliases": {
      "get": {
        "tags": [
          "Aliases"
        ],
        "summary": "Get all aliases",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/counters": {
      "get": {
        "tags": [
          "Counters"
        ],
        "summary": "Get all counters",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/ftl/getallqueries": {
      "get": {
        "tags": [
          "FTL"
        ],
        "summary": "Get all queries that FTL has in memory",
        "parameters": [
          {
            "in": "query",
            "name": "getAllQueries",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/messages": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Get all messages",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/networks": {
      "get": {
        "tags": [
          "Networks"
        ],
        "summary": "Get all networks",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/addresses": {
      "get": {
        "tags": [
          "Network addresses"
        ],
        "summary": "Get all network addresses",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/queries": {
      "get": {
        "tags": [
          "Queries"
        ],
        "summary": "Get all queries",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "X-API-KEY",
        "in": "header"
      }
    }
  }
}